MAINT: move plugin tests to separate file
